+++ pkg/api/mlflow/dao/repositories/param.go
@@ -6,7 +6,7 @@ import (
 	"github.com/rotisserie/eris"
 	"gorm.io/gorm"
 
-	"github.com/G-Research/fasttrackml/pkg/models"
+	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
 )
 
 // ParamRepositoryProvider provides an interface to work with models.Param entity.
