+++ pkg/api/mlflow/dao/repositories/run.go
@@ -9,8 +9,8 @@ import (
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 
+	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
 	"github.com/G-Research/fasttrackml/pkg/database"
-	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 // RunRepositoryProvider provides an interface to work with models.Run entity.
