+++ pkg/api/mlflow/dao/repositories/tag.go
@@ -7,8 +7,8 @@ import (
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
 
+	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
 	"github.com/G-Research/fasttrackml/pkg/database"
-	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 // TagRepositoryProvider provides an interface to work with models.Tag entity.
